Use longs instead of ints, Rational in macros

+++ core/src/main/scala/unu/number/RationalValue.scala
@@ -2,7 +2,7 @@ package unu.number
 
 import scala.language.experimental.macros
 
-case class RationalValue[R](num: Int, denom: Int)
+case class RationalValue[R](num: Long, denom: Long)
 
 object RationalValue {
   implicit def fromMacro[R <: Rational]: RationalValue[R] = macro unu.Macro.materializeRationalValue[R]
